Move integration tests to more appropriate locations
